Data, Code, and SQL output examples
This .py code pulls data from excel workbooks with multiple spreadsheets where data is filled in by rows.  It also clarifies date where month and year are pulled from the file name and day is pulled from the sheet name. The code pivots the data to fill in by columns in the SQL Database. Example of file set up and SQL output included.